pr requests&seperate login form comp. in loginpage
