+++ ecommerce/frontend/src/components/LoginForm.less
@@ -1,0 +1,33 @@
+.login-form {
+    display: block;
+    margin-top: 80px;
+    margin-left: auto;
+    margin-right: auto;
+}
+.login-buttons {
+    display: block;
+    justify-content: center;
+    text-align: center;
+}
+
+.login-button {
+    margin-left: 8px;
+    margin-right: 8px;
+    margin-top: 4px;
+    margin-bottom: 4px;
+}
+
+.router-buttons {
+    display: block;
+    justify-content: center;
+    margin-top: 50px;
+    text-align: center;
+}
+
+.google-login-button {
+   margin: 10px;
+}
+
+.signup-button {
+    margin: 20px;
+}
