--- ecommerce/frontend/src/components/pages/LoginPage.less
@@ -1,34 +1,0 @@
-.loginForm {
-    display: block;
-    margin-top: 80px;
-    margin-left: auto;
-    margin-right: auto;
-    width: 40%;
-}
-.loginButtons {
-    display: block;
-    justify-content: center;
-    text-align: center;
-}
-
-.loginButton {
-    margin-left: 8px;
-    margin-right: 8px;
-    margin-top: 4px;
-    margin-bottom: 4px;
-}
-
-.routerButtons {
-    display: block;
-    justify-content: center;
-    margin-top: 50px;
-    text-align: center;
-}
-
-.googleLoginButton {
-   margin: 10px;
-}
-
-.signUpButton {
-    margin: 20px;
-}
